workers/faust/pyproject.toml
[tool.poetry]
name = "python_worker"
version = "0.1.0"
description = ""
authors = ["Jeremy Nelissen <jeremy@blackcape.io>"]

[tool.poetry.dependencies]
python = "^3.8"
faust = "^1.10.4"
kafka-python = "<2.0.0"
minio = "^6.0.0"
pydantic = "^1.6.1"
toml = "^0.10.1"
urllib3 = "^1.25.10"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
isort = "<5"
mypy = "^0.782"
pylint = "^2.5.3"
pytest = "^5.2"
coverage = "^5.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
